Moves delegation logic to blockchain explorer
