+++ app/frontend/wallet/blockchain-explorer.ts
@@ -3,6 +3,15 @@ import range from './helpers/range'
 import NamedError from '../helpers/NamedError'
 import debugLog from '../helpers/debugLog'
 import getHash from '../helpers/getHash'
+import {
+  DelegetionHistoryItemType,
+  RewardWithdrawal,
+  StakeDelegation,
+  StakePool,
+  StakingReward,
+} from '../components/pages/delegations/delegationHistoryPage'
+import distinct from '../helpers/distinct'
+import {UNKNOWN_POOL_NAME} from './constants'
 
 const cacheResults = (maxAge: number, cache_obj: Object = {}) => <T extends Function>(fn: T): T => {
   const wrapped = (...args) => {
@@ -170,7 +179,19 @@ const blockchainExplorer = (ADALITE_CONFIG) => {
     })
   }
 
-  async function getDelegationHistory(accountPubkeyHex) {
+  async function getPoolInfo(url) {
+    const response = await request(
+      `${ADALITE_CONFIG.ADALITE_SERVER_URL}/api/poolMeta`,
+      'POST',
+      JSON.stringify({poolUrl: url}),
+      {'Content-Type': 'application/json'}
+    ).catch(() => {
+      return {}
+    })
+    return response
+  }
+
+  async function getDelegationHistory(accountPubkeyHex, validStakepools) {
     // Urls
     const delegationsUrl = `https://explorer-staging.adalite.io/api/account/delegationHistory/${accountPubkeyHex}`
     // const rewardsUrl = `https://explorer-staging.adalite.io/api/account/rewardHistory/${accountPubkeyHex}`
@@ -186,7 +207,86 @@ const blockchainExplorer = (ADALITE_CONFIG) => {
     const rewards = [] //await rewardsRequest
     const withdrawals = await withdrawalsRequest
 
-    return {delegations, rewards, withdrawals}
+    const extractUrl = (poolHash) =>
+      validStakepools[poolHash] ? validStakepools[poolHash].url : null
+
+    const poolMetaUrls = distinct(
+      [...delegations, ...rewards].map(({poolHash}) => extractUrl(poolHash))
+    ).filter((x) => x != null)
+
+    // Run requests for meta data in parallel
+    const metaDataPromises = poolMetaUrls.map((url: string) => ({
+      url,
+      metaDataPromise: getPoolInfo(url),
+    }))
+    const metaUrlToPoolNameMap = {}
+    for (const promise of metaDataPromises) {
+      const metaData = await promise.metaDataPromise
+      metaUrlToPoolNameMap[promise.url] = metaData.name
+    }
+
+    const poolHashToPoolName = (poolHash) => {
+      const poolName = metaUrlToPoolNameMap[extractUrl(poolHash)]
+      return poolName || UNKNOWN_POOL_NAME
+    }
+
+    const parseStakePool = (delegationHistoryObject) => {
+      const stakePool: StakePool = {
+        id: delegationHistoryObject.poolHash,
+        name: poolHashToPoolName(delegationHistoryObject.poolHash),
+      }
+
+      return stakePool
+    }
+
+    // Prepare delegations
+    let oldPool: StakePool = null
+    const parsedDelegations = delegations
+      .map((delegation) => ({...delegation, time: new Date(delegation.time)}))
+      .sort((a, b) => a.time.getTime() - b.time.getTime()) // sort by time, oldest first
+      .map((delegation) => {
+        const stakePool: StakePool = parseStakePool(delegation)
+        const stakeDelegation: StakeDelegation = {
+          type: DelegetionHistoryItemType.StakeDelegation,
+          epoch: delegation.epochNo,
+          dateTime: new Date(delegation.time),
+          newStakePool: stakePool,
+          oldStakePool: oldPool,
+        }
+        oldPool = stakePool
+
+        return stakeDelegation
+      })
+      .reverse() // newest first
+
+    // Prepare rewards
+    const parsedRewards = rewards.map((reward) => {
+      const stakingReward: StakingReward = {
+        type: DelegetionHistoryItemType.StakingReward,
+        epoch: reward.epochNo,
+        dateTime: new Date(reward.time),
+        reward: reward.amount,
+        stakePool: parseStakePool(reward),
+      }
+
+      return stakingReward
+    })
+
+    // Prepare withdrawals
+    const parsedWithdrawals = withdrawals.map((withdrawal) => {
+      const rewardWithdrawal: RewardWithdrawal = {
+        type: DelegetionHistoryItemType.RewardWithdrawal,
+        epoch: withdrawal.epochNo,
+        dateTime: new Date(withdrawal.time),
+        credit: withdrawal.amount,
+      }
+
+      return rewardWithdrawal
+    })
+
+    return [...parsedDelegations, ...parsedRewards, ...parsedWithdrawals].sort(
+      (a, b) => b.dateTime.getTime() - a.dateTime.getTime()
+    ) // sort by time, newest first
   }
 
   return {
@@ -198,6 +298,7 @@ const blockchainExplorer = (ADALITE_CONFIG) => {
     getBalance,
     fetchTxInfo,
     filterUsedAddresses,
+    getPoolInfo,
     getDelegationHistory,
   }
 }
